refactor(dispatch): isolate runtime detection scheduling
